% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRIHomology.R
\name{LRI_homology_conversion}
\alias{LRI_homology_conversion}
\title{LRIHomology conversion}
\usage{
LRI_homology_conversion(source_file, inTax, outTax, output_file)
}
\arguments{
\item{source_file}{The file path of the source ligand-receptor data.}

\item{output_file}{The file path to save the converted data.}

\item{source_tax}{The Taxonomy ID of the source species (e.g., 9606 for human).}

\item{target_tax}{The Taxonomy ID of the target species.}
}
\value{
A CSV file with the converted homology data.
}
\description{
This package can convert ligand and receptor data related to human or mouse development into data for different species.
Convert Homology Data for Ligand and Receptor Analysis
}
