% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTFTHomology.R
\name{RTFT_homology_conversion}
\alias{RTFT_homology_conversion}
\title{RTFTHomology conversion}
\usage{
RTFT_homology_conversion(source_file, inTax, outTax, output_file)
}
\arguments{
\item{source_file}{The file path of the source R-TF/TF-T data.}

\item{output_file}{The file path to save the converted data.}

\item{source_tax}{The Taxonomy ID of the source species (e.g., 9606 for human).}

\item{target_tax}{The Taxonomy ID of the target species.}
}
\value{
A CSV file with the converted homology data.
}
\description{
This package provides homology analysis of human and mouse receptors, transcription factors, and target genes.
}
